Cambios menores para Codacy
