+++ pom.xml
@@ -50,7 +50,7 @@
 		<dependency>
 			<groupId>org.postgresql</groupId>
 			<artifactId>postgresql</artifactId>
-			<version>42.6.0</version>
+			<version>42.6.1</version>
 		</dependency>
 		<dependency>
 			<groupId>org.projectlombok</groupId>
