+++ src/main/java/com/rolesync/rolesync/dto/profilecontroller/ProfileInDTO.java
@@ -1,4 +1,4 @@
-package com.rolesync.rolesync.dto.profileControllerDTOs;
+package com.rolesync.rolesync.dto.profilecontroller;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
